remove seteuid call
Originally I was planning to have this run as a chained command,
so this would start as root (via SETUID bit), make the needed
directories, drop privileges, exec into more programs.

Realized it's probably safer to just have this run as a sub-process
of the init scripts and NOT do an exec chain - that way if I mess
something up with regard to UID/EUID/GID/etc behavior, I can't launch
the entire process tree as root.

But - I left the seteuid call in the initial commit. woops!